service/requirements.txt
torch
numpy==1.26.4
supervision==0.18.0
opencv-python
opencv-python-headless
faiss-cpu==1.10.0
sentence-transformers==3.4.1
fastapi[standard]
ultralytics
easyocr
paddlepaddle
paddleocr
timm
einops==0.8.0
ollama==0.4.7
watchdog
python-multipart
google-genai
python-dotenv